Improved the way incorrectly clustered points are counted
Earlier for a found cluster to be considered a match with a known cluster
the first point of the found cluster should be clustered correctly,
or the whole match is missed.

Now for each known cluster the cluster that matches it best
in terms of similar points is found, allowing for the
match to be found even if the first point is clustered incorrectly.